pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "torax"
description = "Differentiable 1D tokamak plasma transport simulator in JAX."
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [{name = "TORAX authors", email="torax-team@google.com"}]
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Intended Audience :: Science/Research",
]

dependencies = [
    "absl-py>=2.0.0",
    "typing_extensions>=4.2.0",
    "typeguard>=2.13.3,<3",
    "immutabledict>=1.0.0",
    "jax>=0.4.32",
    "jaxlib>=0.4.32",
    "jaxopt>=0.8.2",
    "flax>=0.10.0",
    "matplotlib>=3.3.0",
    "numpy>2",
    "setuptools;python_version>='3.10'",
    "chex>=0.1.88",
    "equinox>=0.11.3",
    "PyYAML>=6.0.1",
    "xarray>=2024.11.0",
    "netcdf4>=1.7.2",
    "h5netcdf>=1.3.0",
    "scipy>=1.13.0",
    "jaxtyping>=0.2.28",
    "contourpy>=1.2.1",
    "eqdsk>=0.4.0",
    "pydantic>=2.10.5",
    "tqdm>=4.67.0",
    "treelib>=1.3.2",
]

[project.urls]
homepage = "https://github.com/google-deepmind/torax"
repository = "https://github.com/google-deepmind/torax"
documentation = "https://github.com/google-deepmind/torax/blob/main/README.md"

[project.optional-dependencies]
# Installed through `pip install -e .[dev]`
dev = [
    "pytest",
    "pytest-xdist",
    "pytest-shard",
    "pylint>=2.6.0",
    "pyink",
    "sphinxcontrib-apidoc",
    "sphinx-autodoc-typehints",
    "sphinx-rtd-theme",
    "sphinx_copybutton",
    "sphinx_design",
    "sphinx_remove_toctrees",
]
# Installed through `pip install -e .[tutorial]`
tutorial = [
    "seaborn",
    "notebook",
]

imas = [
    "imas>=4.0.0",
    "imaspy",
]

[tool.pyink]
# Formatting configuration to follow Google style-guide
line-length = 80
unstable = true
pyink-indentation = 2
pyink-use-majority-quotes = true

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["torax*"]

[tool.pytest.ini_options]
addopts = [
    "--ignore=torax/tests/test_data",
    "--ignore=torax/tests/scripts",
    "--ignore-glob=*lib*",
    # Leads to an ImportError in collection for `pytest` without qualikiz_tools.
    "--ignore=torax/transport_model/qualikiz_wrapper.py",
]
testpaths = "**/tests"
python_files = "*.py"
